meus_projetos/prompt-engineering-task/requirements.txt
argparse
requests
beautifulsoup4
python-pptx
cohere